requirement.txt
numpy==1.21.0
flask
nltk
torch
mysql-connector-python